# fix: keep order of fields with `BaseModel.construct()`