--- a/changes/2281-PrettyWood.md
+++ b/changes/2281-PrettyWood.md
@@ -1,0 +1,1 @@
+fix: keep order of fields with `BaseModel.construct()`