<!-- Thank you for your contribution! -->
<!-- Unless your change is trivial, please create an issue to discuss the change before creating a PR -->
<!-- See https://pydantic-docs.helpmanual.io/contributing/ for help on Contributing -->

## Change Summary
Order of fields with `BaseModel.construct()` was not maintained with the current implementation.
I changed it a little. It doesn't seem to have a bad impact of performance (which is the whole purpose of `construct`) on the contrary

```python
from pydantic import BaseModel


class Foo(BaseModel):
    a: int
    b: int = 42
    c: float


def f():
    # Foo.construct(**{'a': 1, 'b': 42, 'c': 3.14})
    Foo.construct(**{'a': 1, 'c': 3.14})


if __name__ == '__main__':
    import timeit
    print(timeit.timeit("f()", number=10_000_000, setup="from __main__ import f"))
    # {'a': 1, 'b': 42, 'c': 3.14}
    # BEFORE: 28.9s
    # AFTER: 23.3s

    # {'a': 1, 'c': 3.14}
    # BEFORE: 27.5s
    # AFTER: 26.6s
```

## Related issue number
closes #2281

## Checklist

* [ ] Unit tests for the changes exist
* [ ] Tests pass on CI and coverage remains at 100%
* [ ] Documentation reflects the changes where applicable
* [ ] `changes/<pull request or issue id>-<github username>.md` file added describing change
  (see [changes/README.md](https://github.com/samuelcolvin/pydantic/blob/master/changes/README.md) for details)