@layday 2021-01-24
This changes the behaviour of `construct` in that it requires the `values` to be in the model's `__fields__`, i.e. it used to behave as if `extra` was set to `allow`.